consumer/requirements.txt
psycopg2-binary
kafka-python
requests
prometheus_client